Fix pull-request requirement and js file organizations: - Add directory js_assets - Add home.js file - Add contact.js file - Change tab file name to menu, with only menu module - Move initPageLoad (alias home) module to home file - Move contact module to contact file - Move erase module to init-pg-ld file Working Ok
