+++ src/js_assets/menu.js
@@ -1,28 +1,5 @@
 /* eslint-disable no-unused-vars, no-eval  */
-import { crtDOM } from './init-pg-ld';
-
-// ----- Module for Erase-Tab content -----
-const erase = (() => {
-  const elem = (e) => { e.remove(); };
-  const pointedElm = (tag) => {
-    const domElem = document.querySelector(tag);
-    erase.elem(domElem);
-  };
-
-  return { elem, pointedElm };
-})();
-
-// ----- Module for Contact-Tab content -----
-const contact = (() => {
-  const makeContactTab = () => {
-    erase.pointedElm('#headline');
-    const headline = crtDOM('div').mkChildNattribute('#content', 'id', 'headline');
-    const hlH2 = crtDOM('h2').mkChildNattNtext('#headline', 'class', 'headline-h', 'Hungry? We\'re here for you!');
-    const hlP = crtDOM('p').mkChildNattNtext('#headline', 'class', 'headline-p', 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.');
-  };
-
-  return { makeContactTab };
-})();
+import { crtDOM, erase } from './init-pg-ld';
 
 // ----- Module for Menu-Tab content -----
 const menu = (() => {
@@ -71,4 +48,4 @@ const menu = (() => {
   };
 })();
 
-export { erase, contact, menu };
+export default menu;
